fix comments make booking id from long to UUID fix conflict and fix compiling proccess

+++ src/main/java/com/spring/booking/dto/request/HotelRequest.java
@@ -10,7 +10,7 @@ public record HotelRequest(
         String name,
         String description,
         String address,
-        BigDecimal rating,
+        Float rating,
         LocalTime checkInTime,
         LocalTime checkOutTime,
         Set<Long> attachmentIds,
